adding swagger support

+++ care/emr/api/viewsets/facility.py
@@ -23,6 +23,7 @@ from care.emr.resources.user.spec import UserSpec
 from care.facility.models import Facility
 from care.security.authorization import AuthorizationController
 from care.users.models import User
+from care.utils.decorators.schema_decorator import generate_swagger_schema_decorator
 from care.utils.file_uploads.cover_image import delete_cover_image, upload_cover_image
 from care.utils.models.validators import (
     cover_image_validator,
@@ -71,6 +72,7 @@ class FacilityFilters(filters.FilterSet):
     phone_number = CharFilter(field_name="phone_number", lookup_expr="iexact")
 
 
+@generate_swagger_schema_decorator
 class FacilityViewSet(EMRModelViewSet):
     database_model = Facility
     pydantic_model = FacilityCreateSpec
@@ -112,25 +114,25 @@ class FacilityViewSet(EMRModelViewSet):
             raise PermissionDenied("Only Super Admins can delete Facilities")
 
     @method_decorator(parser_classes([MultiPartParser]))
-    @action(methods=["POST"], detail=True)
+    @action(methods=["POST", "DELETE"], detail=True)
     def cover_image(self, request, external_id):
         facility = self.get_object()
         self.authorize_update({}, facility)
-        serializer = FacilityImageUploadSerializer(facility, data=request.data)
-        serializer.is_valid(raise_exception=True)
-        serializer.save()
-        return Response(serializer.data)
 
-    @cover_image.mapping.delete
-    def cover_image_delete(self, *args, **kwargs):
-        facility = self.get_object()
-        self.authorize_update({}, facility)
-        delete_cover_image(facility.cover_image_url, "cover_images")
-        facility.cover_image_url = None
-        facility.save()
-        return Response(status=204)
+        if request.method == "POST":
+            serializer = FacilityImageUploadSerializer(facility, data=request.data)
+            serializer.is_valid(raise_exception=True)
+            serializer.save()
+            return Response(serializer.data)
+        if request.method == "DELETE":
+            delete_cover_image(facility.cover_image_url, "cover_images")
+            facility.cover_image_url = None
+            facility.save()
+            return Response(status=204)
+        return Response(data="Method Not Allowed", status=405)
 
 
+@generate_swagger_schema_decorator
 class FacilitySchedulableUsersViewSet(EMRModelReadOnlyViewSet):
     database_model = User
     pydantic_read_model = UserSpec
@@ -149,6 +151,7 @@ class FacilityUserFilter(FilterSet):
     username = CharFilter(field_name="username", lookup_expr="icontains")
 
 
+@generate_swagger_schema_decorator
 class FacilityUsersViewSet(EMRModelReadOnlyViewSet):
     database_model = User
     pydantic_read_model = UserSpec
@@ -163,6 +166,7 @@ class FacilityUsersViewSet(EMRModelReadOnlyViewSet):
         )
 
 
+@generate_swagger_schema_decorator
 class AllFacilityViewSet(EMRModelReadOnlyViewSet):
     permission_classes = ()
     authentication_classes = ()
