+++ care/emr/api/viewsets/user.py
@@ -22,6 +22,7 @@ from care.security.authorization import AuthorizationController
 from care.security.models import RoleModel
 from care.users.api.serializers.user import UserImageUploadSerializer, UserSerializer
 from care.users.models import User
+from care.utils.decorators.schema_decorator import generate_swagger_schema_decorator
 from care.utils.file_uploads.cover_image import delete_cover_image
 
 
@@ -34,6 +35,7 @@ class UserFilter(filters.FilterSet):
     user_type = filters.CharFilter(field_name="username", lookup_expr="iexact")
 
 
+@generate_swagger_schema_decorator
 class UserViewSet(EMRModelViewSet):
     database_model = User
     pydantic_model = UserCreateSpec
@@ -98,25 +100,25 @@ class UserViewSet(EMRModelViewSet):
         return Response(status=200)
 
     @method_decorator(parser_classes([MultiPartParser]))
-    @action(detail=True, methods=["POST"], permission_classes=[IsAuthenticated])
+    @action(
+        detail=True, methods=["POST", "DELETE"], permission_classes=[IsAuthenticated]
+    )
     def profile_picture(self, request, *args, **kwargs):
         user = self.get_object()
         if not self.authorize_update({}, user):
             raise PermissionDenied("Permission Denied")
-        serializer = UserImageUploadSerializer(user, data=request.data)
-        serializer.is_valid(raise_exception=True)
-        serializer.save()
-        return Response(status=200)
 
-    @profile_picture.mapping.delete
-    def profile_picture_delete(self, request, *args, **kwargs):
-        user = self.get_object()
-        if not self.authorize_update({}, user):
-            raise PermissionDenied("Permission Denied")
-        delete_cover_image(user.profile_picture_url, "avatars")
-        user.profile_picture_url = None
-        user.save()
-        return Response(status=204)
+        if request.method == "POST":
+            serializer = UserImageUploadSerializer(user, data=request.data)
+            serializer.is_valid(raise_exception=True)
+            serializer.save()
+            return Response(status=200)
+        if request.method == "DELETE":
+            delete_cover_image(user.profile_picture_url, "avatars")
+            user.profile_picture_url = None
+            user.save()
+            return Response(status=204)
+        return Response(data="Method Not Allowed", status=405)
 
     @action(
         detail=True,
